move contirbutors to metadata store
